Move the daemon recovery regression test to Swift Testing
