+++ cmuxTests/WorkspaceRemoteConnectionTests.swift
@@ -3269,54 +3269,6 @@ final class WorkspaceRemoteConnectionTests: XCTestCase {
         )
     }
 
-    @MainActor
-    func testRecoveredDaemonTransportBounceClearsSidebarDaemonError() {
-        let workspace = Workspace()
-        let config = WorkspaceRemoteConfiguration(
-            destination: "dev@example.com",
-            port: 22,
-            identityFile: nil,
-            sshOptions: [],
-            localProxyPort: nil,
-            relayPort: 64_021,
-            relayID: String(repeating: "c", count: 16),
-            relayToken: String(repeating: "d", count: 64),
-            localSocketPath: "/tmp/cmux-debug-test.sock",
-            terminalStartupCommand: "ssh dev@example.com",
-            preserveAfterTerminalExit: true,
-            skipDaemonBootstrap: true
-        )
-        workspace.configureRemoteConnection(config, autoConnect: false)
-
-        workspace.applyRemoteDaemonStatusUpdate(
-            WorkspaceRemoteDaemonStatus(state: .ready),
-            target: "dev@example.com"
-        )
-        workspace.applyRemoteDaemonStatusUpdate(
-            WorkspaceRemoteDaemonStatus(
-                state: .error,
-                detail: "Remote daemon transport needs re-bootstrap after proxy failure (retry 1 in 2s)"
-            ),
-            target: "dev@example.com"
-        )
-
-        XCTAssertEqual(workspace.logEntries.last?.level, .error)
-
-        workspace.applyRemoteDaemonStatusUpdate(
-            WorkspaceRemoteDaemonStatus(state: .ready),
-            target: "dev@example.com"
-        )
-
-        XCTAssertNil(
-            workspace.logEntries.last(where: { $0.source == "remote-daemon" }),
-            "A recovered daemon transport bounce must retract its sidebar error"
-        )
-        XCTAssertNotEqual(
-            workspace.logEntries.last?.level,
-            .error,
-            "The workspace row must not keep rendering the recovered daemon error"
-        )
-    }
 }
 
 final class CLINotifyProcessIntegrationTests: XCTestCase {
